% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inchToFeet.R
\name{inchToFeet}
\alias{inchToFeet}
\title{inchToFeet}
\usage{
inchToFeet(inches)
}
\arguments{
\item{inches}{Numerical vector, an inch value to be transformed to feet and inch.}
}
\description{
Transforms inches to feet and remaining inches
}
\author{
joschkacr
}
