% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnitConv.R
\name{UnitConv}
\alias{UnitConv}
\title{Conversion function}
\usage{
UnitConv(value, from = c("inch", "cm", "pounds", "kg"), to = c("inch",
  "pounds", "cm", "kg"), r = 2)
}
\arguments{
\item{value}{The value to be converted.}

\item{from}{Unit of given value.}

\item{to}{Unit in which value will be converted.}

\item{r}{Number of decimals}
}
\description{
This function converts values from one unit (inch, cm, pounds or kg) to another unit.
}
\author{
Kafi
}
